VeriFast can't avoid FreeBSD-SA-20:03.thrmisc
https://groups.google.com/g/verifast/c/vJUViRAQbkI/m/1uSJJ-VOBAAJ

Bart said:
"VeriFast currently does not help you find this type of problems.
It currently treats an uninitialized variable just like any other
variable. This is not strictly compliant with the C standard, which
states that if the program reads an uninitialized variable whose
address is never taken, its behavior is undefined (C18 §6.3.2.1).
(Interestingly, from this paragraph it seems that reading an
uninitialized variable whose address is taken is not a problem.)
So it would make sense for a future version of VeriFast to distinguish
between initialized and uninitialized variables and not allow reading
uninitialized variables."